% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accident_data.R
\name{as_accident_data}
\alias{as_accident_data}
\title{Convert a list to an accident_data object}
\usage{
as_accident_data(x)
}
\arguments{
\item{x}{A list containing accident data. The list must contain followings:
\code{accident_info}, \code{person_info}, and \code{highway_info}.}
}
\value{
A \code{accident_data} object.
}
\description{
\code{as_accident_data} converts a list to an \code{accident_data} object.
}
\keyword{internal}
