% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-deg.R
\name{convert_deg}
\alias{convert_deg}
\title{Convert DMS (degree-minute-second) to decimal degrees}
\usage{
convert_deg(dms)
}
\arguments{
\item{dms}{A numeric vector of DMS values in the format of "DDDMMSSsss"}
}
\value{
A numeric vector of decimal degrees.
}
\description{
This function converts DMS (degree-minute-second) to decimal degrees.
The input DMS values must be in the format of "DDDMMSSsss" where:
\itemize{
\item D: Degrees (1--3 digits)
\item M: Minutes (2 digits)
\item S: Seconds (2 digits)
\item s: Milliseconds (3 digits)
}
}
\examples{
convert_deg(c("431007628", "1410328320"))
convert_deg(c(431007628, 1410328320))
}
