% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-dataset-name.R
\name{detect_dataset_name}
\alias{detect_dataset_name}
\title{Identify dataset names from file paths}
\usage{
detect_dataset_name(files, quiet = FALSE)
}
\arguments{
\item{files}{A character string specifying the file paths.}

\item{quiet}{A logical indicating whether to suppress warnings.}
}
\value{
A character vector of dataset names
}
\description{
This function identifies dataset names based on the file paths and
configuration patterns.
}
\details{
Invalid files are replaced with \code{NA}, and a warning is issued.
}
\examples{
detect_dataset_name(c("example/honhyo_2022.csv", "example/honhyo_2023.csv"))
}
