% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-accident-data.R
\name{download_accident_data}
\alias{download_accident_data}
\title{Download the traffic accident data}
\usage{
download_accident_data(type, download_dir = getwd(), years = 2023)
}
\arguments{
\item{type}{A character string specifying the type of data to download:
one of "main", "supp", or "highway".}

\item{download_dir}{A character string specifying the directory to save the
downloaded files (default: \code{getwd()}).}

\item{years}{An integer or vector of integers specifying the years of the
data to download. Available years are 2019 to 2023 (default: \code{2023}).}
}
\value{
A character vector of file paths for the downloaded files
(invisibly).
}
\description{
\code{download_accident_data()} downloads the file of traffic accident data
provided the \href{https://www.npa.go.jp/publications/statistics/koutsuu/opendata/index_opendata.html}{National Police Agency}.
}
\details{
The traffic accident data are divided into the following types:
\itemize{
\item \code{"main"}    : Basic data on traffic accidents (honhyo).
\item \code{"supp"}    : Additional data related to the \code{"main"} data (hojuhyo).
\item \code{"highway"} : Data on traffic accidents on highways (kosokuhyo).
}

Use the \code{type} argument to specify which data type to download.
}
\examples{
\dontrun{
# Download the main data for 2023
download_accident_data("main", "download-dir-path", 2023)

# Download the supplementary data for 2023
download_accident_data("supp", "download-dir-path", 2023)

# Download the highway data for 2019 to 2023
download_accident_data("highway", "download-dir-path", 2019:2023)
}
}
