% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-accident-data.R
\name{filter_accident_data}
\alias{filter_accident_data}
\title{Filter accident data}
\usage{
filter_accident_data(data, filter_keys)
}
\arguments{
\item{data}{A accident_data object.}

\item{filter_keys}{A data frame containing key columns to filter the data.}
}
\value{
A filtered accident_data object.
}
\description{
\code{filter_accident_data()} filters accident data, retaining only the records
that have matching values in the key columns specified in \code{filter_keys}.
}
\keyword{internal}
