% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-accident-type.R
\name{filter_accident_type}
\alias{filter_accident_type}
\title{Filter accident data by accident type}
\usage{
filter_accident_type(data, accident_type)
}
\arguments{
\item{data}{A list containing accident data.}

\item{accident_type}{A vector of specifying the accident type to
filter. Must be one of the following values:
\itemize{
\item \code{"vehicle_to_pedestrian"}: Vehicle-to-pedestrian collision
\item \code{"vehicle_to_vehicle"}   : Vehicle-to-vehicle collision
\item \code{"single_vehicle"}       : Single-vehicle accident
\item \code{"train"}                : Train accident
}}
}
\value{
A list containing filtered accident data.
}
\description{
\code{filter_accident_type()} filters accident data based on the accident type.
}
\examples{
\dontrun{
# Filter accident data by vehicle-to-pedestrian collisions
vehicle_to_pedestrian_collisions <- filter_accident_type(
  data,
  "vehicle_to_pedestrian"
)

# Filter accident data by vehicle-to-vehicle collisions
vehicle_to_pedestrian_collisions <- filter_accident_type(
  data,
  "vehicle_to_vehicle"
)
}
}
