% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-accident-data.R
\name{filter_by_category}
\alias{filter_by_category}
\title{Filter accident data by category}
\usage{
filter_by_category(data, category, category_name)
}
\arguments{
\item{data}{A \code{accident_data} object.}

\item{category}{A vector specifying the category values to filter by.}

\item{category_name}{A string specifying the name of the category column.}
}
\value{
A filtered \code{accident_data} object.
}
\description{
\code{filter_by_category()} filters accident data based on the specified category
and its values.
}
\keyword{internal}
