% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-accident-data.R
\name{filter_by_key}
\alias{filter_by_key}
\title{Filter accident data by matching keys}
\usage{
filter_by_key(data, filter_keys)
}
\arguments{
\item{data}{A \code{accident_data} object.}

\item{filter_keys}{A data frame containing keys for filtering.}
}
\value{
A filtered \code{accident_data} object.
}
\description{
\code{filter_by_key()} filters accident data based on a set of matching keys.
}
\keyword{internal}
