% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-collision-pattern.R
\name{filter_collision_pattern}
\alias{filter_collision_pattern}
\title{Filter collision patterns from accident data}
\usage{
filter_collision_pattern(
  data,
  primary_impacts,
  secondary_impacts = NA_character_
)
}
\arguments{
\item{data}{A list containing accident data.}

\item{primary_impacts}{A vector specifying the initial collision points
(primary impact) for each involved party.}

\item{secondary_impacts}{A vector specifying the subsequent collision points
(secondary impact) for each involved party (default: \code{NA_character_}).}
}
\value{
A list containing filtered accident data.
}
\description{
\code{filter_collision_pattern()} filters accident data based on primary and
secondary impact conditions for each party involved in the accident.
}
\details{
The values of \code{primary_impacts} and \code{secondary_impacts} are following:
\itemize{
\item \code{"no_impact"}   : No impact
\item \code{"front"}      : Front of the vehicle
\item \code{"right"}      : Right side of the vehicle
\item \code{"rear"}       : Rear of the vehicle
\item \code{"left"}       : Left side of the vehicle
\item \code{"front_right"}: Front right of the vehicle
\item \code{"rear_right"} : Rear right of the vehicle
\item \code{"rear_left"}  : Rear left of the vehicle
\item \code{"front_left"} : Front left of the vehicle
The order of values in \code{primary_impacts} and \code{secondary_impacts}
corresponds to the party order (\code{party_order}).
}
}
\examples{
\dontrun{
# Filter accident data by cases indicating head-on collisions
head_on_collisions <- filter_collision_pattern(data, c("front", "front"))

# Filter accident data by cases indicating rear-end collisions
rear_end_collisions <- filter_collision_pattern(data, c("rear", "front"))
}
}
