% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-data-by-tag.R
\name{filter_data_by_tag}
\alias{filter_data_by_tag}
\title{Filter data by tag}
\usage{
filter_data_by_tag(data, tag)
}
\arguments{
\item{data}{An accident data object.}

\item{tag}{A character string specifying the tag to filter by.}
}
\value{
A data frame containing only the columns with the specified tag.
}
\description{
This function filters data by a specific tag.
}
\examples{
\dontrun{
person_data <- filter_data_by_tag(accident_data, "person")
}
}
