% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-dataset-name.R
\name{find_dataset_name}
\alias{find_dataset_name}
\title{Find dataset name from file path}
\usage{
find_dataset_name(file)
}
\arguments{
\item{file}{A character string specifying the file path.}
}
\value{
Dataset name
}
\description{
This function determines the dataset name based on the file path and
configuration.
}
\examples{
find_dataset_name("example/honhyo_2023.csv")
}
