% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Get configuration}
\usage{
get_config()
}
\value{
A list containing the configuration.
}
\description{
This function retrieves the configuration that were loaded using
\code{load_config()}.
}
\examples{
\dontrun{
config <- get_config()
}
}
\seealso{
\code{\link[=load_config]{load_config()}} for loading the configuration.

\code{\link[=reset_config]{reset_config()}} for resetting the configuration.
}
