% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-default-config-path.R
\name{get_default_config_path}
\alias{get_default_config_path}
\title{Get the path to the default configuration file}
\usage{
get_default_config_path()
}
\value{
The path to the default configuration file
}
\description{
This function retrieves the path to the default configuration YAML file for
the jpaccidents package.
}
\examples{
get_default_config_path()
}
