% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpaccidents-package.R
\docType{package}
\name{jpaccidents-package}
\alias{jpaccidents}
\alias{jpaccidents-package}
\title{jpaccidents: Tools for Traffic Accident Data in Japan}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp}

}
\keyword{internal}
