% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-column-names.R
\name{lookup_column_names}
\alias{lookup_column_names}
\title{Lookup new column names for an accident dataset}
\usage{
lookup_column_names(dataset_name, original_names)
}
\arguments{
\item{dataset_name}{The name of the dataset.}

\item{original_names}{A vector of original column names.}
}
\value{
A vector of new column names corresponding to the provided
original names.
}
\description{
This function retreieves the new column names corresponding to
the original column names in a specified dataset. The column
names are based on the configuration loaded by \code{load_config()}.
This is useful when the original names are in Japanese (2-byte
characters), which can be difficult to handle in programs.
This function converts them to corresponding English names.
}
\examples{
lookup_column_names("main_data", c("死者数", "天候"))
}
