% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{post_process}
\alias{post_process}
\alias{post_process.accident_data}
\alias{post_process_main}
\alias{post_process_sub}
\alias{post_process_highway}
\title{Post-process data}
\usage{
post_process(data)

\method{post_process}{accident_data}(data)

post_process_main(data)

post_process_sub(data)

post_process_highway(data)
}
\arguments{
\item{data}{The data to be processed.}
}
\value{
A processed dataset as a list.
}
\description{
\code{post_process()} performs post-processing on different datasets:
\itemize{
\item \code{post_process_main()} handles main data (honhyo) by separating it into
accident and person data.
\item \code{post_process_sub()} handles sub data (hojuhyo), processing
person-related information.
\item \code{post_process_highway()} handles highway data (kosokuhyo), extracting
highway-related information.
}
}
\details{
These functions process the data based on its type, ensuring consistency
and proper formatting.
}
