% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{read_accident_file}
\alias{read_accident_file}
\title{Read CSV file with common parameters}
\usage{
read_accident_file(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file}
}
\value{
tibble with raw accident data
}
\description{
Read CSV file with common parameters
}
\keyword{internal}
