% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{read_accidents_data}
\alias{read_accidents_data}
\title{Read accident data from a CSV file}
\usage{
read_accidents_data(file_paths)
}
\arguments{
\item{file_paths}{A character vector of file paths to the CSV files.}
}
\value{
A list containing the accident data split into three data frames:
\code{accidents_info}, \code{persons_info}, and \code{highways_info}.
}
\description{
This function reads accident data from multiple CSV files,
processes each file, and combines the data into three categories:
\itemize{
\item \code{accidents_info}
\item \code{persons_info}
\item \code{highways_info}
}
}
\examples{
\dontrun{
read_accidents_data(c(
  "example/honhyo_2023.csv",
  "example/hojuhyo_2023.csv"
))
}
}
