% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{reset_config}
\alias{reset_config}
\title{Reset Configuration}
\usage{
reset_config()
}
\description{
This function resets the configuration to the default values.
}
\examples{
reset_config()
}
\seealso{
\code{\link[=get_config]{get_config()}} for retrieving the configuration.

\code{\link[=load_config]{load_config()}} for loading the configuration.
}
