% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{select_post_processor}
\alias{select_post_processor}
\title{Select post-processor function based on dataset name}
\usage{
select_post_processor(dataset_name)
}
\arguments{
\item{dataset_name}{The name of the dataset.}
}
\value{
The post-processing function corresponding to the dataset name.
}
\description{
This function selects the post-processor function based on
the given dataset name.
}
