% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{tidy_accident_data}
\alias{tidy_accident_data}
\title{Tidy accident data based on file type}
\usage{
tidy_accident_data(raw_data)
}
\arguments{
\item{raw_data}{Raw data frame with file type attribute}
}
\value{
List of processed data frames
}
\description{
Tidy accident data based on file type
}
\keyword{internal}
