% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{transform_dms_to_decimal}
\alias{transform_dms_to_decimal}
\title{Transform DMS coordinates to decimal degrees}
\usage{
transform_dms_to_decimal(dms)
}
\arguments{
\item{dms}{Numeric vector of coordinates in DDMMSS.sss format}
}
\value{
Numeric vector of decimal degrees
}
\description{
Transform DMS coordinates to decimal degrees
}
\keyword{internal}
