% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{validate_file_paths}
\alias{validate_file_paths}
\title{Validate input file paths}
\usage{
validate_file_paths(file_paths)
}
\arguments{
\item{file_paths}{A character vector of file paths}
}
\value{
A character vector of validated file paths
}
\description{
Validate input file paths
}
\keyword{internal}
