% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-accident-data.R
\name{download_accident_data}
\alias{download_accident_data}
\title{Download the traffic accident data}
\usage{
download_accident_data(type, download_dir = getwd(), years = 2023)
}
\arguments{
\item{type}{Type of data to download: one of "main", "sub", or "highway".}

\item{download_dir}{A directory where downloaded file is to saved
(default: \code{getwd()}).}

\item{years}{Years the traffic accident data were recorded.
Available from 2019 to 2023. You can specify a single year or
vector of years(default: \code{2023}).}
}
\value{
Path of the downloaded file (invisibly).
}
\description{
Download the file of traffic accident data provided
\href{https://www.npa.go.jp/publications/statistics/koutsuu/opendata/index_opendata.html}{the National Police Agency's Web page}.
}
\details{
The traffic accident data are divided into the following types:
\itemize{
\item "main" (honhyo): Basic data on traffic accidents.
\item "sub" (hojuhyo): Additional data related to the "main" data.
\item "highway" (kosokuhyo): Data on traffic accidents on highways.
}

You can specify the type of data directly via the \code{type} argument.
}
\examples{
\dontrun{
download_accident_data("main", "download-dir-path", 2023)
download_accident_data("sub", "download-dir-path", 2023)
download_accident_data("highway", "download-dir-path", 2023)
}
}
