% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpaccidents-package.R
\docType{package}
\name{jpaccidents-package}
\alias{jpaccidents}
\alias{jpaccidents-package}
\title{jpaccidents: Tools for Traffic Accident Data in Japan}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

In Japan, traffic accident data are published by the National Police Agency. This package provides tools to efficiently handle and standardise these datasets, facilitating seamless analysis and integration into various workflows. By using this package, users can uncover patterns in traffic accidents and identify high-risk scenarios, supporting efforts to enhance road safety and prevent future incidents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nononoexe.github.io/jpaccidents/}
}

}
\author{
\strong{Maintainer}: Keisuke Ando \email{ando@maslab.aitech.ac.jp} (\href{https://orcid.org/0009-0002-4085-2067}{ORCID})

Other contributors:
\itemize{
  \item Takeshi Uchitane [thesis advisor]
  \item Naoto Mukai [thesis advisor]
  \item Kazunori Iwata [thesis advisor]
  \item Nobuhiro Ito [thesis advisor]
}

}
\keyword{internal}
