% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpaccidents-package.R
\docType{package}
\name{jpaccidents-package}
\alias{jpaccidents}
\alias{jpaccidents-package}
\title{jpaccidents: Tools for Traffic Accident Data in Japan}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nononoexe.github.io/jpaccidents/}
}

}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp}

}
\keyword{internal}
