% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-process.R
\name{post_process}
\alias{post_process}
\alias{post_process.accident_data}
\alias{post_process_main}
\alias{post_process_sub}
\alias{post_process_highway}
\title{Post-process data}
\usage{
post_process(data)

\method{post_process}{accident_data}(data)

post_process_main(data)

post_process_sub(data)

post_process_highway(data)
}
\arguments{
\item{data}{The data to be processed.}
}
\value{
A processed data.
}
\description{
These function perform post-processing operations on different datasets.
\code{post_process_main} processes the main data (honhyo), separating it into
accident and person data, and returning them as a list.
\code{post_process_sub} and \code{post_process_highway} currently do not perform
any modifications on sub data and highway data, respectively.
}
