% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{read_accident_data}
\alias{read_accident_data}
\alias{read_accident_main_data}
\alias{read_accident_sub_data}
\alias{read_accident_highway_data}
\title{Read the traffic accident data}
\usage{
read_accident_main_data(file)

read_accident_sub_data(file)

read_accident_highway_data(file)
}
\arguments{
\item{file}{A character string of the file path.}
}
\value{
\code{read_accident_main_data()} separates traffic accident data
and injured party data from the target data and returns them as a list.
\code{read_accident_sub_data()} and \code{read_accident_highway_data()} return
the contents of the target data as a tibble.
}
\description{
Read the file of traffic accident data.

Use the appropriate function for the type of data you want to read:
\itemize{
\item \code{read_accident_main_data()} to read the main data (honhyo).
\item \code{read_accident_sub_data()} to read the sub data (hojuhyo).
\item \code{read_accident_highway_data()} to read the highway data (kosokuhyo).
}
}
\examples{
\dontrun{
read_accident_main_data("downloaded-file-path.csv")
read_accident_sub_data("downloaded-file-path.csv")
read_accident_highway_data("downloaded-file-path.csv")
}
}
