% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{read_accident_data}
\alias{read_accident_data}
\title{Read accident data from a CSV file}
\usage{
read_accident_data(
  file,
  post_process = select_post_processor(find_dataset_name(file))
)
}
\arguments{
\item{file}{The path to the CSV file.}

\item{post_process}{A function to apply post-processing to the data.}
}
\value{
A data frame containing the accident data.
}
\description{
This function reads accident data from a CSV file,
renames the columns, adds attributes and class, and applies
post-processing if provided.
}
\examples{
\dontrun{
read_accident_data("example/honhyo_2023.csv")
}
}
