% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-accident-data.R
\name{read_accident_data}
\alias{read_accident_data}
\title{Read accident data from CSV files}
\usage{
read_accident_data(file_paths)
}
\arguments{
\item{file_paths}{A character vector of file paths to the CSV files.}
}
\value{
A list containing three data frames (\code{accident_info}, \code{person_info},
and \code{highway_info}).
}
\description{
\code{read_accident_data()} reads accident data from multiple CSV files,
processes each file, and combines the data into three categories:
\itemize{
\item \code{accident_info}: Basic accident information including date, location,
and conditions
\item \code{person_info}  : Information about individuals involved in the accidents
\item \code{highway_info} : Information specific to accidents that occurred on
highways/expressways
}
}
\examples{
\dontrun{
# Read main and supplementary data for 2023
read_accident_data(c("example/honhyo_2023.csv", "example/hojuhyo_2023.csv"))
}
}
